refactor: 입력 값에 대해 유효성 검사를 위한 클래스 분리
